Componentized instructor
