+++ src/css/Signup.css
@@ -4,7 +4,6 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    padding: 24px;
   }
   
   .content {
